PB/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.0
django-phonenumber-field==7.0.0
Pillow==9.3.0
django-google-maps==0.13.0
geopy==2.2.0
requests==2.28.1
django-eventtools==1.0.3
django-nested-admin==4.0.0
django-credit-cards==0.4.1